fix connector TYS url and add test
